requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai>=0.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
email-validator==2.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
itsdangerous==2.1.2
Jinja2==3.1.2
blinker==1.6.2
